审批流迁移

See merge request hand-rongjing-hec/hcf-front!265

+++ src/containers/setting/workflow/widget/widget.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { connect } from 'dva'
 import { Icon, Popconfirm, message } from 'antd'
-import PropTypes from 'prop-types';
 
 import manApprovalImg from 'images/setting/workflow/man-approval.svg'
 import knowImg from 'images/setting/workflow/know.svg'
@@ -9,6 +8,7 @@ import aiApprovalImg from 'images/setting/workflow/aiapproval.svg'
 import mailImg from 'images/setting/workflow/mail.png'
 import endImg from 'images/setting/workflow/end.png'
 import auditImg from 'images/setting/workflow/audit.png'
+import PropTypes from 'prop-types';
 
 import workflowService from 'containers/setting/workflow/workflow.service'
 import 'styles/setting/workflow/widget/widget.scss'
@@ -37,17 +37,22 @@ class Widget extends React.Component{
 
   //删除节点
   handleDelete = (e, widget) => {
-    workflowService.deleteApprovalNode(widget.ruleApprovalNodeOID).then(() => {
-      message.success(this.$t('common.delete.success', {name: ''}));
-      this.props.deleteHandle(true)
-    })
+    if (this.props.isRuleInEdit) {
+      message.warning(this.$t('setting.key1319'/*你有一个编辑中的审批条件未保存*/));
+    } else {
+      this.props.deleteHandle();
+      workflowService.deleteApprovalNode(widget.ruleApprovalNodeOID).then(() => {
+        message.success(this.$t('common.delete.success', {name: ''}));
+        this.props.deleteHandle(true)
+      })
+    }
   };
 
   renderWidget = (widget) => {
     return (
       <div className={`widget-list-item ${this.props.className}`}>
         <div className="top-cover"/>
-        <Popconfirm onConfirm={e => this.handleDelete(e, widget)} title={this.$t('workflow.detail.confirm.delete.node')/*是否确认删除该节点？*/}>
+        <Popconfirm onConfirm={e => this.handleDelete(e, widget)} title={this.$t('setting.key1411'/*是否确认删除该节点？*/)}>
           <Icon type="close-circle" className="delete-icon" onClick={e => {e.preventDefault();e.stopPropagation()}}/>
         </Popconfirm>
         <img src={this.getImageUrl(widget.type)}/>
@@ -71,6 +76,7 @@ Widget.propTypes = {
   widget: PropTypes.object,
   width: PropTypes.any,
   className: PropTypes.string,
+  isRuleInEdit: PropTypes.bool,
   deleteHandle: PropTypes.func,
 };
 
@@ -86,4 +92,4 @@ function mapStateToProps(state) {
   }
 }
 
-export default connect(mapStateToProps, null, null, { withRef: true })(Widget)
+export default connect(mapStateToProps)(Widget)
